exercises 1-10

+++ misc-programs/1_5_2_character_counting_2.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    long nc;
+    double nc;
     for (nc = 0; getchar() != EOF; nc++) {
         /* Do Nothing */
     }
 
-    printf("%ld\n", nc);
+    printf("%f\n", nc);
 }
